github-graphql-sample/pyproject.toml
[project]
name = "github-graphql-sample"
version = "0.1.0"
description = "A command-line application demonstrating how to use GitHub's GraphQL API with personal access token authentication"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "GitHub Copilot", email = "noreply@github.com"},
]
keywords = ["github", "graphql", "api", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "azure-identity>=1.13.0",
]

[project.scripts]
github-graphql = "github_graphql_client:main"

[dependency-groups]
dev = [
    "black>=24.8.0",
    "flake8>=5.0.4",
    "pytest>=8.3.5",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
